chore: move validation helpers to utils
