+++ src/__tests__/validateTypes.unit.test.ts
@@ -1,47 +1,8 @@
-import validateTypes, { typeOf, hasSameTypes } from '../lib/validateTypes'
+import validateTypes from '../lib/validateTypes'
 
 import { ddbDocClient as DocumentClient } from './bootstrap-tests'
 
 describe('validateTypes', () => {
-  describe('typeOf', () => {
-    it('detects null type', () => {
-      expect(typeOf(null)).toBe('null')
-    })
-
-    it('detects binary type', () => {
-      expect(typeOf(new Int32Array())).toBe('Binary')
-    })
-
-    it('detects string type', () => {
-      expect(typeOf('str')).toBe('String')
-    })
-
-    it('detects object type', () => {
-      expect(typeOf(Object.create(null))).toBe('Object')
-    })
-
-    it('detects undefined type', () => {
-      expect(typeOf()).toBe('undefined')
-    })
-  })
-
-  describe('hasSameType', () => {
-    it('should have same types for empty array', () => {
-      expect(hasSameTypes([])).toBeTruthy()
-    })
-
-    it('should have same types for one element array', () => {
-      expect(hasSameTypes([''])).toBeTruthy()
-    })
-
-    it('should have same types for all numbers array', () => {
-      expect(hasSameTypes([1, 2])).toBeTruthy()
-    })
-
-    it('should have different types for string/numbers array', () => {
-      expect(hasSameTypes([1, '2'])).toBeFalsy()
-    })
-  })
 
   it('validates string', async () => {
     let result = validateTypes(DocumentClient)(
